requirements.txt
fastapi
uvicorn
itchat-uos==1.5.0.dev0
pydantic
requests
